Move setup files to directory misc and rename ERD to ERM
