+++ Vagrantfile
@@ -64,5 +64,5 @@ Vagrant.configure(2) do |config|
   # Enable provisioning with a shell script. Additional provisioners such as
   # Puppet, Chef, Ansible, Salt, and Docker are also available. Please see the
   # documentation for more information about their specific syntax and use.
-  config.vm.provision "shell", path: "provision.sh" 
+  config.vm.provision "shell", path: "misc/provision.sh" 
 end
