+++ datajoint/user_relations.py
@@ -2,11 +2,9 @@
 Hosts the table tiers, user relations should be derived from.
 """
 
-import re
-import abc
 from datajoint.relation import Relation
 from .autopopulate import AutoPopulate
-from datajoint.utils import from_camel_case
+from utils import from_camel_case
 
 
 class Manual(Relation):
@@ -98,19 +96,3 @@ class Subordinate:
         raise NotImplementedError('Subtables should not be populated directly.')
 
 
-def from_camel_case(s):
-    """
-    Convert names in camel case into underscore (_) separated names
-
-    Example:
-    >>>from_camel_case("TableName")
-        "table_name"
-    """
-
-    def convert(match):
-        return ('_' if match.groups()[0] else '') + match.group(0).lower()
-
-    if not re.match(r'[A-Z][a-zA-Z0-9]*', s):
-        raise DataJointError(
-            'ClassName must be alphanumeric in CamelCase, begin with a capital letter')
-    return re.sub(r'(\B[A-Z])|(\b[A-Z])', convert, s)
