+++ tests/test_utils.py
@@ -2,8 +2,9 @@
 Collection of test cases to test core module.
 """
 from nose.tools import assert_true, assert_raises, assert_equal
-from datajoint.user_relations import from_camel_case
 from datajoint import DataJointError
+from utils import from_camel_case
+
 
 def setup():
     pass
